algoritmo implementando, falta corrigir bugs e os testes

+++ PCV/genetico/main.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
     ### Input ###
     args, graph = fileinput()
 
-    genetic(graph[:], args.population, args.mutation, args.max_iteration, args.max_time * 60, args.seed)
+    genetic(graph[:], args.population, args.mutation, int(args.max_iteration), int(args.max_time * 60) , args.seed)
